requirement.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.9
elasticsearch==7.13.0
idna==2.10
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
PySocks==1.7.1
requests==2.25.1
supervisor==4.2.2
traitlets==5.0.5
ujson==4.0.2
urllib3==1.26.5
uvloop==0.15.2
wcwidth==0.2.5
websocket-client==1.0.1
